+++ index.cjs
@@ -1,3 +1,4 @@
 const { postnumer } = require("./postnumer.cjs");
+const { thjodskraPlaces } = require("./places.cjs");
 
-module.exports = { postnumer };
+module.exports = { postnumer, thjodskraPlaces };