+++ index.mjs
@@ -1,3 +1,4 @@
 import { postnumer } from "./postnumer.mjs";
+import { thjodskraPlaces } from "./places.mjs";
 
-export { postnumer };
+export { postnumer, thjodskraPlaces };